requirements.txt
tensorflow
keras
numpy
pillow
nltk
kagglehub
